requirements.txt
aenum==3.0.0
alabaster==0.7.12
anytree==2.8.0
appdirs==1.4.4
appnope==0.1.2
argon2-cffi==20.1.0
arrow==0.17.0
astroid==2.5.1
async-generator==1.10
attrs==20.3.0
Babel==2.9.0
backcall==0.2.0
backports.entry-points-selectable==1.1.0
bleach==3.3.1
blinker==1.4
breathe==4.27.0
cachetools==4.2.4
canopen==1.2.1
cbor==1.0.0
certifi==2021.5.30
cffi==1.14.5
chardet==4.0.0
charset-normalizer==2.0.6
click==7.1.2
cmsis-pack-manager==0.2.10
colorama==0.4.4
coverage==5.5
cryptography==3.4.6
cycler==0.10.0
decorator==5.0.9
defusedxml==0.7.1
distlib==0.3.3
Django==2.2
django-admin-charts==0.24.1
django-bower==5.2.0
django-jsonfield==1.4.1
django-nvd3==0.9.7
django-simple-history==3.0.0
djangorestframework==3.9.2
djcacheutils==3.0.0
docopt==0.6.2
docutils==0.16
entrypoints==0.3
filelock==3.2.0
Flask==1.1.2
Flask-Cors==3.0.10
Flask-Mail==0.9.1
future==0.18.2
gcovr==4.2
gitlint==0.15.0
google-api-core==2.1.0
google-auth==2.2.1
google-cloud-bigquery==2.28.0
google-cloud-core==2.1.0
google-crc32c==1.3.0
google-resumable-media==2.0.3
googleapis-common-protos==1.53.0
grpcio==1.41.0
gunicorn==20.1.0
hidapi==0.10.1
hypothesis==6.10.1
idna==3.2
imagesize==1.2.0
imgtool==1.7.1
importlib-metadata==3.7.0
importlib-resources==5.2.2
iniconfig==1.1.1
intelhex==2.3.0
intervaltree==3.1.0
ipykernel==5.3.2
ipython==7.16.1
ipython-genutils==0.2.0
isort==5.6.4
itsdangerous==1.1.0
jedi==0.18.0
Jinja2==2.11.2
jsonschema==3.2.0
junit2html==30.0.6
junitparser==1.6.3
jupyter-client==6.2.0
jupyter-core==4.7.1
jupyterlab-pygments==0.1.2
kiwisolver==1.3.1
lazy-object-proxy==1.4.3
lpc-checksum==2.2.0
lxml==4.6.2
MarkupSafe==1.1.1
matplotlib==3.2.2
mccabe==0.6.1
milksnake==0.1.5
mistune==0.8.4
mock==4.0.3
mypy==0.812
mypy-extensions==0.4.3
naturalsort==1.5.1
nbclient==0.5.3
nbconvert==6.0.7
nbformat==5.1.3
nest-asyncio==1.5.1
notebook==6.4.0
numpy==1.19.5
opencv-contrib-python==3.4.2.17
opencv-python==3.4.2.17
packaging==21.0
pandocfilters==1.4.3
parso==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.1.1
pipenv==2021.5.29
platformdirs==2.4.0
pluggy==0.13.1
ply==3.11
prettytable==2.1.0
progress==1.5
prometheus-client==0.11.0
prompt-toolkit==3.0.19
proto-plus==1.19.2
protobuf==3.18.1
psutil==5.8.0
psycopg2==2.7.1
psycopg2-binary==2.9.1
ptyprocess==0.7.0
py==1.10.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pyelftools==0.27
Pygments==2.8.0
PyJWT==2.0.1
pykwalify==1.8.0
pylink-square==0.8.2
pylint==2.7.2
pyocd==0.29.0
pyparsing==2.4.7
pyrsistent==0.18.0
pyserial==3.5
pytest==6.2.2
python-can==3.3.4
python-dateutil==2.8.1
python-magic==0.4.22
python-memcached==1.59
python-nvd3==0.14.2
python-slugify==1.1.4
pytz==2021.3
pyusb==1.1.1
PyYAML==5.4.1
pyzmq==22.1.0
requests==2.26.0
rsa==4.7.2
ruamel.yaml==0.16.12
ruamel.yaml.clib==0.2.2
Send2Trash==1.7.1
sh==1.14.1
six==1.16.0
snowballstemmer==2.1.0
sortedcontainers==2.3.0
Sphinx==3.3.1
sphinx-rtd-theme==0.5.1
sphinx-tabs==2.0.1
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==1.0.3
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.4
sphinxcontrib-svg2pdfconverter==1.1.1
sqlparse==0.4.2
tabulate==0.8.9
terminado==0.10.1
testpath==0.5.0
toml==0.10.2
tornado==6.1
traitlets==4.3.3
typed-ast==1.4.1
typing-extensions==3.7.4.3
Unidecode==1.3.2
urllib3==1.26.7
virtualenv==20.8.1
virtualenv-clone==0.5.7
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==1.0.1
west==0.9.0
whitenoise==5.3.0
wrapt==1.12.1
zipp==3.4.0
